+++ CHANGELOG.md
@@ -9,12 +9,13 @@
 * add comments
 * clj-reload
 * :chatgpt is unnecessary. if found chatgpt's answer, show it.
-* revice answers.clj
-* digest
+* digest namespace
+
 
 # 0.2.8-SNAPSHOT
 
-* valid=>status
+- changed valid => status (true/false => "yes"/"no")
+- changed admin/upsert!
 
 # 0.2.7 (2025-09-15)